Moved errors to router.rs
