+++ rust-runtime/aws-smithy-http-server/src/proto/aws_json/aws_json_10/mod.rs
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-pub mod error;
+pub mod router;
 
 /// [AWS JSON 1.0 Protocol](https://awslabs.github.io/smithy/2.0/aws/protocols/aws-json-1_0-protocol.html).
 pub struct AwsJson10;
