+++ rust-runtime/aws-smithy-http-server/src/proto/aws_json/aws_json_10/router.rs
@@ -5,19 +5,19 @@
 
 use crate::body::{empty, BoxBody};
 use crate::extension::RuntimeErrorExtension;
-use crate::proto::aws_json::error::Error;
+use crate::proto::aws_json::router::Error;
 use crate::response::IntoResponse;
 use crate::routers::{method_disallowed, UNKNOWN_OPERATION_EXCEPTION};
 
-use super::AwsJson11;
+use super::AwsJson10;
 
-impl IntoResponse<AwsJson11> for Error {
+impl IntoResponse<AwsJson10> for Error {
     fn into_response(self) -> http::Response<BoxBody> {
         match self {
             Error::MethodNotAllowed => method_disallowed(),
             _ => http::Response::builder()
                 .status(http::StatusCode::NOT_FOUND)
-                .header(http::header::CONTENT_TYPE, "application/x-amz-json-1.1")
+                .header(http::header::CONTENT_TYPE, "application/x-amz-json-1.0")
                 .extension(RuntimeErrorExtension::new(
                     UNKNOWN_OPERATION_EXCEPTION.to_string(),
                 ))
