+++ rust-runtime/aws-smithy-http-server/src/proto/rest/rest_json_1/mod.rs
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-pub mod error;
+pub mod router;
 
 /// [AWS REST JSON 1.0 Protocol](https://awslabs.github.io/smithy/2.0/aws/protocols/aws-restjson1-protocol.html).
 pub struct AwsRestJson1;
