+++ rust-runtime/aws-smithy-http-server/src/proto/rest/rest_json_1/router.rs
@@ -3,27 +3,25 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use crate::body::empty;
 use crate::body::BoxBody;
 use crate::extension::RuntimeErrorExtension;
-use crate::proto::rest::error::Error;
+use crate::proto::rest::router::Error;
 use crate::response::IntoResponse;
-use crate::routers::method_disallowed;
-use crate::routers::UNKNOWN_OPERATION_EXCEPTION;
+use crate::routers::{method_disallowed, UNKNOWN_OPERATION_EXCEPTION};
 
-use super::AwsRestXml;
+use super::AwsRestJson1;
 
-/// An AWS REST routing error.
-impl IntoResponse<AwsRestXml> for Error {
+impl IntoResponse<AwsRestJson1> for Error {
     fn into_response(self) -> http::Response<BoxBody> {
         match self {
             Error::NotFound => http::Response::builder()
                 .status(http::StatusCode::NOT_FOUND)
-                .header(http::header::CONTENT_TYPE, "application/xml")
+                .header(http::header::CONTENT_TYPE, "application/json")
+                .header("X-Amzn-Errortype", UNKNOWN_OPERATION_EXCEPTION)
                 .extension(RuntimeErrorExtension::new(
                     UNKNOWN_OPERATION_EXCEPTION.to_string(),
                 ))
-                .body(empty())
+                .body(crate::body::to_boxed("{}"))
                 .expect("invalid HTTP response for REST JSON routing error; please file a bug report under https://github.com/awslabs/smithy-rs/issues"),
             Error::MethodNotAllowed => method_disallowed(),
         }
