+++ rust-runtime/aws-smithy-http-server/src/proto/rest/rest_xml/mod.rs
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-pub mod error;
+pub mod router;
 
 /// [AWS REST XML Protocol](https://awslabs.github.io/smithy/2.0/aws/protocols/aws-restxml-protocol.html).
 pub struct AwsRestXml;
